neochat_django/requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
django-cors-headers>=4.9,<4.10
channels>=4.3,<4.4
daphne>=4.2,<4.3
requests>=2.32,<2.33
channels-redis>=4.2,<4.3
psycopg[binary]>=3.2,<3.3
python-dotenv>=1.0,<1.1
